Get rid of the special-case of the /build/ directory.
We wanted it for the (auto-generated) build directories in the
live-editor subrepo, but now we have a build directory in the top
level of webapp as well.  We could make these regexps more specific,
so they apply only to live-editor, but I think that's fragile (what if
we add a "css" subdir to webapp/build?)  So I'm just taking them out
for now.

Auditors: john

Test plan:
Fingers crossed